Affichage pieuvre

+++ tripletface.cpp
@@ -1,42 +1,90 @@
 #include "tripletface.h"
 #include <sstream>
 
+/**
+ * Constructeur TripletFace
+ * @brief TripletFace::TripletFace
+ * @param v
+ * @param vt
+ * @param vn
+ */
 TripletFace::TripletFace(int v, int vt, int vn){
     this->v=v;
     this->vt=vt;
     this->vn=vn;
 }
 
+/**
+ * Constructeur TripletFace
+ * @brief TripletFace::TripletFace
+ * @param v
+ * @param vt
+ */
 TripletFace::TripletFace(int v, int vt){
     this->v=v;
     this->vt=vt;
     vn=-1;
 }
 
-double TripletFace::getV(){
+/**
+ * Getter index vertex
+ * @brief TripletFace::getV
+ * @return index Vertex
+ */
+int TripletFace::getV(){
     return v;
 }
 
-double TripletFace::getVT(){
+/**
+ * Getteer index texture
+ * @brief TripletFace::getVT
+ * @return index texture
+ */
+int TripletFace::getVT(){
     return vt;
 }
 
-double TripletFace::getVN(){
+/**
+ * Getter index normale
+ * @brief TripletFace::getVN
+ * @return index normale
+ */
+int TripletFace::getVN(){
     return vn;
 }
 
-void TripletFace::setV(double v){
+/**
+ * Setter index vertex
+ * @brief TripletFace::setV
+ * @param v
+ */
+void TripletFace::setV(int v){
     this->v=v;
 }
 
-void TripletFace::setVT(double vt){
+/**
+ * Setter index texture
+ * @brief TripletFace::setVT
+ * @param vt
+ */
+void TripletFace::setVT(int vt){
     this->vt=vt;
 }
 
-void TripletFace::setVN(double vn){
+/**
+ * Setter index normale
+ * @brief TripletFace::setVN
+ * @param vn
+ */
+void TripletFace::setVN(int vn){
     this->vn=vn;
 }
 
+/**
+ * Impression du triplet "v1/t1/n1"
+ * @brief TripletFace::print
+ * @return le triplet au format obj
+ */
 string TripletFace::print(){
     stringstream ligne;
     if(vn!=-1){
